requirements.txt
streamlit==1.51.0
bcrypt==5.0.0
SQLAlchemy==2.0.44
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0
plotly==6.4.0
